services/encryption/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask-Testing==0.7.1
requests==2.21.0
flask-cors==3.0.7